Create README.md
Note:  Electronic nanostructures  are divided   into four categories based on their dimensions, which are in the nanoscale: zero-dimensional, one-dimensional, two-dimensional nanomaterials, and  three-dimensional bulk nanostructures  , which can be produced by top-down or bottom-up methods. Due to their very small size, nanostructures   exhibit special properties, sometimes different from other conventional materials, in electrical conductivity.
In general,  nanostructures   have three dimensions: length, width, and height. If at least one of these dimensions is on the nanotechnology scale (1-100 nanometers), the material is called a nanostructure.  Nanostructures   are divided into different categories depending on how many dimensions they have on the nanotechnology scale. One of these categories is based on the number of free dimensions. A free dimension is a dimension that is not on the nanoscale and can have any value. Energy structures (levels or bands) of materials exist along each of the length, width, and height dimensions. In other words, every three-dimensional object has three separate energy structures along its three dimensions, the result of which expresses the energy structure of the entire material. Dimensions of nanostructures that are on the nanoscale have so-called quantum confinement. For example, thin layers that have discrete energy levels in one dimension. Quantum confinement means that due to the dimensional constraints at the nanoscale, the energy bands become discrete, and the greater the confinement (smaller the dimensions), the greater the distance between the energy levels. Therefore, one of the main differences between different types of nanostructured materials is the number of continuous energy bands and discrete energy levels in three dimensions, which leads to large variations in their purity.